fstab-generator: fix bind mounts in initrd (issue #6827)
When we mount e.g. /usr in the initrd, using information from
/sysroot/fstab, we prepend /sysroot to the mount directory.  But we also
need to do the same for the source directory of bind mounts.

To reproduce this bug, create an fstab entry `/usr /usr none bind`
and reboot (using systemd in the initrd).  It will fail to load systemd
from the rootfs (ENOENT).  If you interrupt the initrd e.g. pass `rd.break`
to dracut, you can observe (e.g. using findmnt or /proc/self/mountinfo)
that /sysroot/usr was a bind mount of /usr, i.e. the initrd /usr.