fstab-generator: reject mount points which are relative paths
Mount points like "usr" are already rejected with a warning in add_mount().
But we accepted paths like "dev/pts".  So there was something wrong with
this check.

It seems the real intention of the uses of `is_path(where)` was to let
through swap entries like "/dev/sda2 swap swap defaults".  Any other entry
is passed to `mount`, which always interprets `where` as a directory.

This commit rejects all mount points which are relative paths, except for
mount entries which are actually swap entries.